requirements.txt
setuptools
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
httpx
pandas
numpy
matplotlib
pytrends
langchain
langchain_core
langchain_openai
langchain_pinecone
langchain_community
langgraph
pinecone
openai
langfuse
streamlit
azure-identity
azure-keyvault-secrets
plotly
unstructured[local-inference,pdf]
tenacity
tqdm
pillow
google-genai